cleaned up test organization
